+++ analysis/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from .views      import StarView
+
+urlpatterns= [
+    path('star/<int:movieId>', StarView.as_view())
+]
